Split Dyldcache Bin Plugin and Refactor to use RzBinVirtualFile
